feat(catalogo): refatorar estrutura de repositório e interfaces, removendo dependências desnecessárias
