+++ back/microsservicos/catalogo/app/create_catalogo/create_catalogo_usecase.ts
@@ -1,10 +1,11 @@
-import { CatalogoRepositoryMock, CatalogoType } from "../../repo/catalogoRepositoryMock";
+import { Catalogo } from "../../shared/interfaces";
+import { CatalogoRepositoryMock } from "../../shared/repo/catalogoRepositoryMock";
 
 export class CreateCatalogoUsecase {
 
     constructor(private repo: CatalogoRepositoryMock) {}
 
-    public execute(props: CatalogoType) {
+    public execute(props: Catalogo) {
 
         const room = this.repo.createCatalogo(props)
