+++ back/microsservicos/catalogo/app/get_all_catalogo/get_all_catalogo_usecase.ts
@@ -1,4 +1,4 @@
-import { CatalogoRepositoryMock } from "../../repo/catalogoRepositoryMock";
+import { CatalogoRepositoryMock } from "../../shared/repo/catalogoRepositoryMock";
 
 export class GetAllCatalogoUsecase {
