+++ back/microsservicos/catalogo/router.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { getAllCatalogoController } from "./app/get_all_catalogo/get_all_catalogo_presenter";
-import { CatalogoRepositoryMock } from "./repo/catalogoRepositoryMock";
+import { CatalogoRepositoryMock } from "./shared/repo/catalogoRepositoryMock";
 import { createCatalogoController } from "./app/create_catalogo/create_catalogo_presenter";
 import { getCatalogoController } from "./app/get_catalogo/get_catalogo_presenter";
 import { deleteCatalogoController } from "./app/delete_catalogo/delete_catalogo_presenter";
@@ -10,8 +10,8 @@ const router: Router = Router()
 
 router.get('/catalogo', (req, res) => getAllCatalogoController.handle(req, res))
 router.get('/catalogo/:id', (req, res) => getCatalogoController.handle(req, res))
-router.post('/catalogo', async (req, res) => await createCatalogoController.handle(req, res))
-router.delete('/catalogo/:id', async (req, res) => await deleteCatalogoController.handle(req, res))
-router.patch('/catalogo/:id', async (req, res) => await updateCatalogoController.handle(req, res))
+router.post('/catalogo', (req, res) => createCatalogoController.handle(req, res))
+router.delete('/catalogo/:id', (req, res) => deleteCatalogoController.handle(req, res))
+router.patch('/catalogo/:id', (req, res) => updateCatalogoController.handle(req, res))
 
 export { router,  }
