+++ back/microsservicos/catalogo/shared/interfaces.ts
@@ -1,0 +1,10 @@
+export interface Catalogo {
+    id: string
+    name: string
+    description: string
+    address: string
+    comodities: string[]
+    pictures: string[]
+    price: number
+    capacity: number
+}
