+++ back/microsservicos/catalogo/shared/types.ts
@@ -1,0 +1,4 @@
+import { Catalogo } from "./interfaces";
+
+// Deixa id obrigatoria e o resto opcional
+export type updateCatalogoProps = Pick<Catalogo, 'id'> & Partial<Omit<Catalogo, 'id'>>
