+++ back/microsservicos/catalogo/shared/types/Catalogo.ts
@@ -1,0 +1,10 @@
+export type CatalogoType = {
+    id: string
+    name: string
+    description: string
+    address: string
+    comodities: string[]
+    pictures: string[]
+    price: number
+    capacity: number
+}
