+++ back/microsservicos/catalogo/shared/types/Routes.ts
@@ -1,0 +1,1 @@
+import { CatalogoType } from "./Catalogo";
