+++ back/microsservicos/user/shared/interfaces.ts
@@ -1,0 +1,6 @@
+import { authType, informationType } from "./types";
+
+export interface User {
+  auth: authType;
+  information: informationType;
+}
